ultimos cambios

+++ ms-socios/src/main/java/maik/example/mssocios/entity/Socio.java
@@ -15,7 +15,7 @@ public class Socio {
     private String nombre;
     private String apellido;
     private String edad;
-    private Integer dni;
-    private Integer numero;
+    private Double dni;
+    private Double numero;
     private String correo;
 }
